fix: resolve linter and formatting errors after refactor and helper extraction
