+++ .coveragerc
@@ -1,0 +1,3 @@
+[run]
+omit =
+    dgi/models.py
