+++ dgi/models/__init__.py
@@ -1,0 +1,3 @@
+from .company import CompanyData
+
+__all__ = ["CompanyData"]
